# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: OrionOps API
  description: Enterprise Service Orchestration Platform - ITSM + ERP Extensions
  version: 1.0.0
  contact:
    name: OrionOps Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /changes:
    get:
      tags: [Changes]
      summary: List change requests
      operationId: listChanges
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Paginated list
    post:
      tags: [Changes]
      summary: Create change request
      operationId: createChange
      responses:
        '201':
          description: Change created
  /changes/{id}:
    get:
      tags: [Changes]
      summary: Get change request
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Change details
    put:
      tags: [Changes]
      summary: Update change
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Updated
    delete:
      tags: [Changes]
      summary: Delete change
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '204':
          description: Deleted
  /changes/{id}/approve:
    patch:
      tags: [Changes]
      summary: Approve change
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Change approved
  /changes/{id}/reject:
    patch:
      tags: [Changes]
      summary: Reject change
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Change rejected
  /changes/{id}/implement:
    patch:
      tags: [Changes]
      summary: Implement change
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Change implemented
  /changes/{id}/close:
    patch:
      tags: [Changes]
      summary: Close change
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Change closed

  # ---- Service Requests ----
...
  /cmdb/services/{id}:
    get:
      tags: [CMDB]
      summary: Get service
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Service details
    put:
      tags: [CMDB]
      summary: Update service
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Updated
    delete:
      tags: [CMDB]
      summary: Delete service
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '204':
          description: Deleted

  # ---- SLA ----
...
  /audit:
    get:
      tags: [Audit]
      summary: Query audit logs
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - name: entityType
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: string
            format: uuid
        - name: fromDate
          in: query
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Paginated audit logs
  /audit/entity/{type}/{id}:
    get:
      tags: [Audit]
      summary: Get audit trail for entity
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Entity audit trail

  # ---- Search ----
...
  /integrations/{id}:
    get:
      tags: [Integrations]
      summary: Get integration
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Integration details
    put:
      tags: [Integrations]
      summary: Update integration
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Updated
    delete:
      tags: [Integrations]
      summary: Delete integration
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '204':
          description: Deleted
  /integrations/{id}/test:
    post:
      tags: [Integrations]
      summary: Test integration endpoint
      parameters:
        - $ref: '#/components/parameters/EntityIdParam'
      responses:
        '200':
          description: Test result
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 0
      description: Page number (0-indexed)
    SizeParam:
      name: size
      in: query
      schema:
        type: integer
        default: 20
      description: Page size
...
    EntityIdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Entity UUID
...